+++ packages/backend/src/mock/services/email.service.ts
@@ -1,0 +1,9 @@
+const mockEmailService = async () => ({
+  sendJoinEmail: (target: string, uuid: string) =>
+    new Promise((resolve) => resolve({ target, uuid })),
+});
+
+type MockEmailService = Awaited<ReturnType<typeof mockEmailService>>;
+
+export { mockEmailService };
+export type { MockEmailService };
