+++ packages/backend/src/modules/auth/auth.controller.spec.ts
@@ -1,49 +1,27 @@
-import { CreateUserConfirmDTO, CreateUserDTO, User } from '@my-task/common';
-import { BadRequestException } from '@nestjs/common';
+import { User } from '@my-task/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { v4 as uuidv4 } from 'uuid';
+import { MockAuthService, MockEmailService, mockAuthService, mockEmailService } from '~/mock';
 import { EmailService } from '~/modules/email/email.service';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 
 describe('AuthController', () => {
-  let mockAuthService: any;
-  let mockEmailService: any;
+  let authService: MockAuthService;
+  let emailService: MockEmailService;
   let controller: AuthController;
 
   beforeEach(async () => {
-    mockAuthService = {
-      uuidToEmail: new Map<string, CreateUserDTO>(),
-      emailToUuid: new Map<string, string>(),
-      createUser(dto: CreateUserDTO) {
-        const uuid = uuidv4();
-        this.uuidToEmail.set(uuid, dto);
-        this.emailToUuid.set(dto.email, uuid);
-        return uuid;
-      },
-      createUserConfirm(dto: CreateUserConfirmDTO) {
-        if (!this.uuidToEmail.has(dto.uuid))
-          throw new BadRequestException('UUID cannot be found: Wrong DTO!');
-        const data = this.uuidToEmail.get(dto.uuid) as CreateUserDTO;
-
-        const user: User = { id: Math.floor(Math.random() * 10), ...data };
-        return user;
-      },
-    };
-
-    mockEmailService = {
-      sendJoinEmail: (target: string, uuid: string) =>
-        new Promise((resolve) => resolve({ target, uuid })),
-    };
+    [authService, emailService] = await Promise.all([mockAuthService(), mockEmailService()]);
 
     const module: TestingModule = await Test.createTestingModule({
       providers: [EmailService, AuthService],
       controllers: [AuthController],
     })
       .overrideProvider(EmailService)
-      .useValue(mockEmailService)
+      .useValue(emailService)
       .overrideProvider(AuthService)
-      .useValue(mockAuthService)
+      .useValue(authService)
       .compile();
 
     controller = module.get<AuthController>(AuthController);
@@ -72,7 +50,7 @@ describe('AuthController', () => {
       const userToAdd = { email: 'create@example.email' };
       const { email } = controller.createUser(userToAdd);
 
-      const uuid = mockAuthService?.emailToUuid?.get(email);
+      const uuid = authService.emailToUuid.get(email);
       let user: User;
       expect((user = await controller.createUserConfirm({ uuid }))).toBeDefined();
       expect(user.email).toEqual(userToAdd.email);
